Move common object xml parser parts to superclass
